Moves fixtures to be under test/fixtures instead of test/integration.
Removes 'pulp' from the name of all cassettes for consistency.

Fixes some failing tests that are due to inconsistency in to_json of
hashes by adding custom matcher for checking JSON in the body as a
hash comparison.

Adds custom matcher to test against path instead of URI to alleviate
problem with different hostnames for different developers running the
tests.