consumer/group - update to allow for update all
This commit contains a small modification to the generate_content
for consumer and consumer group to enable support for users to
request an 'update all' operation on the pulp interface.  The
basic case where this is used is to request 'update all packages
on consumer X'.  (e.g. equivalent to 'yum update' from the consumer).