fixing runcible errors caused by rubocop changes
when we turned on rubocop, we converted unused method arguments
to start with underscores.  This broke any method that relied on
binding.send(:local_variables).  This disables that cop and fixes
all of those instances.

In addition the tests failed to catch this as vcr cassette matching was not using
body_json as one of its matchers.  This adds body_json as a default matcher.

Once body_json was turned on a ton of tests would fail as repository creation
was using a repo fixture path which included the users home directory.  I changed this
to copy the repo directory to /var/www/ so it will be consistent across users.

In addition I disabled SSL validation as this is not working for tests (nor is it needed)