Fixes #15728 - Updating process_request for json 2.0
With the 2.0 release of the json gem, JSON.parse('true') started to
return true instead of JSON::ParserError (which is technically correct).
However, we send 'body' to RestClient::Response.create which attempts to
extend and modify the body (see https://git.io/vKoeh) which isn't
allowed for true (as it's a frozen value) so instead we now just use the
body ('true') instead of the actual parsed value if the body is not a
Hash or Array.